requirements.txt
py-cord==2.6.1
yt-dlp==2024.8.6
ffmpeg==1.4
python-dotenv==1.0.1
PyNaCl==1.5.0